+++ apps/dine-pleiepenger/src/pages/404.page.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable max-len */
 import type { NextPage } from 'next';
 import { BodyShort, Box, HGrid, Heading, Link, List } from '@navikt/ds-react';
 import { browserEnv } from '../utils/env';