+++ packages/k9-sak-innsyn-k9-sak-api/src/index.ts
@@ -3,3 +3,4 @@ export * from './client/sdk.gen';
 export * from './client/types.gen';
 export * from './client/zod.gen';
 export * from './client/client.gen';
+export * from './initK9SakInnsynK9SakApiClients';