+++ packages/k9-sak-innsyn-k9-sak-api/src/initK9SakInnsynK9SakApiClients.ts
@@ -1,0 +1,7 @@
+import { client } from '.';
+import { initApiClient } from './utils/initApiClient';
+
+export const initK9SakInnsynK9SakApiClients = (config: { frontendPath: string; loginURL: string }) => {
+    const { frontendPath, loginURL } = config;
+    initApiClient(client, frontendPath, loginURL);
+};