+++ packages/sif-common-forms-ds/src/forms/enkeltdatoer/EnkeltdatoListAndDialog.tsx
@@ -2,8 +2,8 @@ import {
     FormikModalFormAndList,
     ModalFormAndListLabels,
     TypedFormInputValidationProps,
+    ValidationError,
 } from '@navikt/sif-common-formik-ds';
-import { ValidationError } from '@navikt/sif-common-formik-ds';
 import { DateRange, sortMaybeDateRange } from '@navikt/sif-common-utils';
 import EnkeltdatoForm from './EnkeltdatoForm';
 import EnkeltdatoList from './EnkeltdatoList';