+++ packages/sif-common-forms-ds/src/forms/fosterbarn/stories/FosterbarnExample.tsx
@@ -1,18 +1,22 @@
 import { Tabs, VStack } from '@navikt/ds-react';
 import { useState } from 'react';
 import { useIntl } from 'react-intl';
-import { getIntlFormErrorHandler, TypedFormikForm, TypedFormikWrapper } from '@navikt/sif-common-formik-ds';
-import { getListValidator } from '@navikt/sif-validation';
-import { ValidationError } from '@navikt/sif-common-formik-ds';
-import { flatten } from 'flat';
-import { fosterbarnMessages } from '..';
 import MessagesPreview from '@navikt/sif-common-core-ds/src/dev-utils/intl/messages-preview/MessagesPreview';
+import {
+    getIntlFormErrorHandler,
+    TypedFormikForm,
+    TypedFormikWrapper,
+    ValidationError,
+} from '@navikt/sif-common-formik-ds';
+import { getListValidator } from '@navikt/sif-validation';
+import { flatten } from 'flat';
+import { fosterbarnMessages } from '../';
+import StoryFormWrapper from '../../../../storybook/components/story-form-wrapper/StoryFormWrapper';
 import SubmitPreview from '../../../../storybook/components/submit-preview/SubmitPreview';
 import FormValidationErrorMessages from '../../../../storybook/components/validation-error-messages/ValidationErrorMessages';
 import FosterbarnForm, { FosterbarnFormErrors } from '../FosterbarnForm';
 import FosterbarnListAndDialog from '../FosterbarnListAndDialog';
 import { Fosterbarn } from '../types';
-import StoryFormWrapper from '../../../../storybook/components/story-form-wrapper/StoryFormWrapper';
 
 enum FormField {
     'fosterbarn' = 'fosterbarn',
@@ -74,7 +78,7 @@ const FormikExample = () => {
                 <Tabs.Panel value="validationMessages">
                     <FormValidationErrorMessages
                         validationErrorIntlKeys={flatten(FosterbarnFormErrors)}
-                        formName={'Fosterbarn'}
+                        formName="Fosterbarn"
                         intlMessages={fosterbarnMessages}
                     />
                 </Tabs.Panel>