+++ packages/sif-common-forms-ds/src/forms/utenlandsk-næring/stories/UtenlandskNæringExample.tsx
@@ -2,17 +2,21 @@ import { Tabs, VStack } from '@navikt/ds-react';
 import { useState } from 'react';
 import { useIntl } from 'react-intl';
 import MessagesPreview from '@navikt/sif-common-core-ds/src/dev-utils/intl/messages-preview/MessagesPreview';
-import { getIntlFormErrorHandler, TypedFormikForm, TypedFormikWrapper } from '@navikt/sif-common-formik-ds';
+import {
+    getIntlFormErrorHandler,
+    TypedFormikForm,
+    TypedFormikWrapper,
+    ValidationError,
+} from '@navikt/sif-common-formik-ds';
 import { getListValidator } from '@navikt/sif-validation';
-import { ValidationError } from '@navikt/sif-common-formik-ds';
 import { flatten } from 'flat';
 import StoryFormWrapper from '../../../../storybook/components/story-form-wrapper/StoryFormWrapper';
 import SubmitPreview from '../../../../storybook/components/submit-preview/SubmitPreview';
 import FormValidationErrorMessages from '../../../../storybook/components/validation-error-messages/ValidationErrorMessages';
+import { utenlandskNæringMessages } from '../i18n';
 import { UtenlandskNæring } from '../types';
 import UtenlandskNæringForm, { UtenlandskNæringFormErrors } from '../UtenlandskNæringForm';
 import UtenlandskNæringListAndDialog from '../UtenlandskNæringListAndDialog';
-import { utenlandskNæringMessages } from '../i18n';
 
 enum FormField {
     'utenlandskNæring' = 'utenlandskNæring',
@@ -81,7 +85,7 @@ const UtenlandskNæringExample = () => {
                 <Tabs.Panel value="validationMessages">
                     <FormValidationErrorMessages
                         validationErrorIntlKeys={flatten(UtenlandskNæringFormErrors)}
-                        formName={'UtenlandskNæring'}
+                        formName="UtenlandskNæring"
                         intlMessages={utenlandskNæringMessages}
                     />
                 </Tabs.Panel>