+++ packages/sif-common-forms-ds/src/forms/virksomhet/VirksomhetInfoAndDialog.tsx
@@ -3,8 +3,8 @@ import {
     FormikModalFormAndInfo,
     ModalFormAndInfoLabels,
     TypedFormInputValidationProps,
+    ValidationError,
 } from '@navikt/sif-common-formik-ds';
-import { ValidationError } from '@navikt/sif-common-formik-ds';
 import { mapVirksomhetToVirksomhetApiData } from './mapVirksomhetToApiData';
 import { Virksomhet } from './types';
 import VirksomhetForm from './VirksomhetForm';