+++ packages/ung-deltakelse-opplyser-api-veileder/src/veileder/client/core/params.ts
@@ -3,18 +3,28 @@ type Slot = 'body' | 'headers' | 'path' | 'query';
 export type Field =
     | {
           in: Exclude<Slot, 'body'>;
+          /**
+           * Field name. This is the name we want the user to see and use.
+           */
           key: string;
+          /**
+           * Field mapped name. This is the name we want to use in the request.
+           * If omitted, we use the same value as `key`.
+           */
           map?: string;
       }
     | {
           in: Extract<Slot, 'body'>;
+          /**
+           * Key isn't required for bodies.
+           */
           key?: string;
           map?: string;
       };
 
 export interface Fields {
     allowExtra?: Partial<Record<Slot, boolean>>;
-    args?: ReadonlyArray<Field>;
+    args?: readonly Field[];
 }
 
 export type FieldsConfig = ReadonlyArray<Field | Fields>;
@@ -71,7 +81,7 @@ const stripEmptySlots = (params: Params) => {
     }
 };
 
-export const buildClientParams = (args: ReadonlyArray<unknown>, fields: FieldsConfig) => {
+export const buildClientParams = (args: readonly unknown[], fields: FieldsConfig) => {
     const params: Params = {
         body: {},
         headers: {},