+++ packages/k9-sak-innsyn-k9-sak-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/k9-brukerdialog-prosessering-api
 
+## 0.0.10
+
+### Patch Changes
+
+- eslint oppsett og endringer for api
+
 ## 0.0.9
 
 ### Patch Changes