+++ packages/ung-common/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @navikt/ung-common
 
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/ung-deltakelse-opplyser-api-deltaker@0.0.19
+    - @navikt/ung-deltakelse-opplyser-api-veileder@0.0.19
+    - @navikt/k9-brukerdialog-prosessering-api@0.0.20
+
 ## 0.0.20
 
 ### Patch Changes